+++ storage/sql/common/rule.go
@@ -69,7 +69,7 @@ func (s *Store) ListRules(ctx context.Context, flagKey string, opts ...storage.Q
 	)
 
 	if params.Limit > 0 {
-		query = query.Limit(uint64(params.Limit) + 1)
+		query = query.Limit(params.Limit + 1)
 	}
 
 	if params.PageToken != "" {
